+++ src/page-objects/AdministrationPages.ts
@@ -34,6 +34,9 @@ import { ManufacturerDetail } from './administration/ManufacturerDetail';
 import { ProductListing } from './administration/ProductListing';
 import { ProductBulkEdit } from './administration/ProductBulkEdit';
 import { CustomerBulkEdit } from './administration/CustomerBulkEdit';
+import { SettingsListing } from './administration/SettingsListing';
+import { DocumentListing } from './administration/DocumentListing';
+import { DocumentDetail } from './administration/DocumentDetail';
 
 export interface AdministrationPageTypes {
     AdminProductDetail: ProductDetail;
@@ -69,6 +72,9 @@ export interface AdministrationPageTypes {
     AdminProductListing: ProductListing;
     AdminProductBulkEdit: ProductBulkEdit;
     AdminCustomerBulkEdit: CustomerBulkEdit;
+    AdminSettingsListing: SettingsListing;
+    AdminDocumentListing: DocumentListing;
+    AdminDocumentDetail: DocumentDetail;
 }
 
 export const AdminPageObjects = {
@@ -105,6 +111,9 @@ export const AdminPageObjects = {
     ProductListing,
     ProductBulkEdit,
     CustomerBulkEdit,
+    SettingsListing,
+    DocumentListing,
+    DocumentDetail,
 }
 
 export const test = base.extend<FixtureTypes>({
@@ -239,5 +248,14 @@ export const test = base.extend<FixtureTypes>({
 
     AdminCustomerBulkEdit: async ({ AdminPage }, use) => {
         await use(new CustomerBulkEdit(AdminPage));
+    },
+    AdminSettingsListing: async ({ AdminPage }, use) => {
+        await use(new SettingsListing(AdminPage));
+    },
+    AdminDocumentListing: async ({ AdminPage }, use) => {
+        await use(new DocumentListing(AdminPage));
+    },
+    AdminDocumentDetail: async ({ AdminPage }, use) => {
+        await use(new DocumentDetail(AdminPage));
     },
 });